+++ src/main/java/rzk/wirelessredstone/client/render/FastTESRFrequency.java
@@ -4,8 +4,11 @@ import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.renderer.BufferBuilder;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraftforge.client.model.animation.FastTESR;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 import rzk.wirelessredstone.tile.TileFrequency;
 
+@SideOnly(Side.CLIENT)
 public class FastTESRFrequency extends FastTESR<TileFrequency>
 {
 	@Override