+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### Version 1.12.2-1.1.0:
+
+- ADDED: Command to clear a frequency from active transmitters
+- ADDED: Wireless Redstone blocks can now be mined faster with a pickaxe
+- ADDED: Cooldown for the remote item (0.5s)
+- FIXED: Potential crashes caused by Remote::onUpdate and BlockFrequency::setPoweredState
+
 ### Version 1.12.2-1.0.0:
 
 - FIXED: Active frequency stays active after transmitter exploded